+++ src/rics/performance/__init__.py
@@ -3,7 +3,7 @@
 from ._format_perf_counter import format_perf_counter, format_seconds
 from ._multi_case_timer import MultiCaseTimer, SkipIfParams
 from ._plot import plot_run
-from ._util import get_best, legacy_plot_run, to_dataframe
+from ._util import get_best, legacy_plot_run, relative_to, to_dataframe
 from ._wrapper import run_multivariate_test
 
 __all__ = [
@@ -14,6 +14,7 @@ __all__ = [
     "get_best",
     "legacy_plot_run",
     "plot_run",
+    "relative_to",
     "run_multivariate_test",
     "to_dataframe",
 ]